requirements.txt
aioredis==1.2.0
alembic==1.0.6
amqp==2.3.2
ansible==2.4.1.0
asgi-redis==1.4.3
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.12.1
Automat==0.7.0
bcrypt==3.1.5
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
cffi==1.11.5
channels==2.1.6
channels-redis==2.3.2
chardet==3.0.4
Click==7.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.4.2
daphne==2.2.4
Django==1.11.6
django-celery-results==1.0.4
django-cors-headers==2.4.0
django-filter==2.0.0
django-redis==4.10.0
django-tinymce==2.6.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
docutils==0.14
drf-extensions==0.4.0
Flask==1.0.2
Flask-Migrate==2.3.1
Flask-Script==2.0.6
Flask-Session==0.3.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
hiredis==0.3.1
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10
kombu==4.2.2.post1
lockfile==0.12.2
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.1.0
msgpack==0.5.6
msgpack-python==0.5.6
paramiko==2.4.2
pid==2.2.1
Pillow==5.3.0
ply==3.11
psutil==5.4.8
pyasn1==0.4.5
pycparser==2.19
pycrypto==2.6.1
pycryptodomex==3.7.2
PyHamcrest==1.9.0
PyJWT==1.7.1
PyMySQL==0.9.3
PyNaCl==1.3.0
pysmi==0.3.3
pysnmp==4.4.8
python-daemon==2.2.0
python-dateutil==2.7.5
python-editor==1.0.3
python-nmap==0.6.1
pytz==2018.9
PyYAML==3.13
redis==3.0.1
requests==2.21.0
service==0.5.1
setproctitle==1.1.10
six==1.12.0
SQLAlchemy==1.2.16
Twisted==18.9.0
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.24.1
vine==1.2.0
Werkzeug==0.14.1
WTForms==2.2.1
zope.interface==4.6.0
